requirements.txt
fastapi==0.85.1
passlib==1.7.4
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.2
starlette==0.20.4
uvicorn==0.15.0
argon2-cffi==21.3.0
python-multipart==0.0.5